requirements.txt
argh
python-slugify
BeautifulSoup4
